refactor: orderTable order와 합병
