+++ src/main/java/kitchenpos/order/infra/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.ordertable.infra;
+package kitchenpos.order.infra;
 
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
